Update careers page, homepage banner, blog list pagination, and refund policy

+++ admin/blog_list.php
@@ -53,10 +53,28 @@ include('sidebar.php');
     <!-- /.content-header -->
 
 <?php
+$limit = 50;
+$page = isset($_GET['page']) && is_numeric($_GET['page']) ? (int)$_GET['page'] : 1;
+$offset = ($page - 1) * $limit;
+$search = isset($_GET['search']) ? mysqli_real_escape_string($link, $_GET['search']) : '';
+
+$where = "WHERE 1";
+if ($search) {
+    $where .= " AND title LIKE '%$search%'";
+}
+
+// Count total published blogs (for the header count)
 $query10 = "SELECT count(*) as rec FROM blog where is_publish='Yes'";
 $result10 = mysqli_query($link, $query10);
 $row10 = $result10->fetch_assoc();
 $publish_blogs = $row10['rec'];
+
+// Count total records for pagination
+$count_query = "SELECT COUNT(*) as total FROM blog $where";
+$count_result = mysqli_query($link, $count_query);
+$count_row = mysqli_fetch_assoc($count_result);
+$total_rows = $count_row['total'];
+$total_pages = ceil($total_rows / $limit);
 ?>
  
  
@@ -70,14 +88,26 @@ $publish_blogs = $row10['rec'];
     </div>
 </div>  
 <br>
+
+<div class="row mb-3" style="width: 70%; margin: 0 auto;">
+    <div class="col-md-12">
+        <form method="GET" action="blog_list.php" class="form-inline justify-content-center">
+            <div class="input-group" style="width: 100%; max-width: 500px;">
+                <input type="text" name="search" class="form-control" placeholder="Search blogs by title..." value="<?php echo htmlspecialchars($search); ?>">
+                <div class="input-group-append">
+                    <button class="btn btn-primary" type="submit" style="background: #3f6791;">Search</button>
+                    <?php if($search): ?>
+                        <a href="blog_list.php" class="btn btn-secondary">Clear</a>
+                    <?php endif; ?>
+                </div>
+            </div>
+        </form>
+    </div>
+</div>
 <br> 
 	  
 	<?php 
-/*$username = "root"; 
-$password = ""; 
-$database = "respo"; 
-$mysqli = new mysqli("localhost", $username, $password, $database); */
-$query = "SELECT * FROM blog order by id desc limit 0,200";
+$query = "SELECT * FROM blog $where ORDER BY id DESC LIMIT $offset, $limit";
 
 echo '<table border="1" cellspacing="2" cellpadding="2" style="width: 70%;" bgcolor="#FFFFF" align="center"> 
       <tr class="tr"> 
@@ -89,18 +119,14 @@ echo '<table border="1" cellspacing="2" cellpadding="2" style="width: 70%;" bgco
 		  <td class="td"> <font face="Arial" color="#FFFFFF">Publish</font> </td> 
 		  <td class="td"> <font face="Arial" color="#FFFFFF">Publish Date</font> </td> 
           <td class="td"> <font face="Arial" color="#FFFFFF">Edit</font> </td>'; 
-?>		  
-<?php
+
 if($_SESSION['user']['username'] == 'finoadmin')
 {
-	  
-		  
 		echo '<td class="td"> <font face="Arial" color="#FFFFFF">Delete</font> </td>'; 
-      
-
 }
   
 echo '</tr>';
+
 if ($result = mysqli_query($link, $query)) {
     while ($row = $result->fetch_assoc()) {
         $field1name = $row["id"];
@@ -113,15 +139,14 @@ if ($result = mysqli_query($link, $query)) {
 		$pd = explode(" ", $created_date);
 		$publish_date = $pd[0];
 		
+		$query1 = "SELECT * FROM blog_category where id = '$field3name'";
+		$result1 = mysqli_query($link, $query1);
+		$row1 = $result1->fetch_assoc();
+		$cat_name = isset($row1['cat_name']) ? $row1['cat_name'] : 'N/A';
 			
-			$query1 = "SELECT * FROM blog_category where id = '$field3name'";
-			$result1 = mysqli_query($link, $query1);
-			$row1 = $result1->fetch_assoc();
-			$cat_name = $row1['cat_name'];
-			
-		if($publish == 'Yes')	
-		{
-        echo '<tr> 
+		$bgcolor = ($publish == 'Yes') ? '' : 'bgcolor="##D3D3D3;"';
+		
+        echo '<tr '.$bgcolor.'> 
                   <td><font color="black">'.$field1name.'</font></td> 
                   <td><font color="black">'.$field2name.'</font></td>
 									  
@@ -130,38 +155,42 @@ if ($result = mysqli_query($link, $query)) {
 				  <td><a href="edit-ipodetail.php?i='.$field1name.'">IPO Detail</td>
 				  <td><font color="black">'.$publish.'</font></td>
 				  <td><font color="black">'.$publish_date.'</font></td>
-                  <td><a href="edit-blog.php?i='.$field1name.'">Edit</td>';
+				  <td><a href="edit-blog.php?i='.$field1name.'">Edit</td>';
 				  if($_SESSION['user']['username'] == 'finoadmin')
 				  {	  
-				  echo '<td><a href="delete-blog.php?i='.$field1name.'">Delete</td>';
+				  echo '<td><a href="delete-blog.php?i='.$field1name.'" onclick="return confirm(\'Are you sure you want to delete this blog?\')">Delete</td>';
 				  }
             echo '</tr>';
-			
-		}
-        else
-        {
-          echo '<tr bgcolor="##D3D3D3;"> 
-                  <td><font color="black">'.$field1name.'</font></td> 
-                  <td><font color="black">'.$field2name.'</font></td>
-									  
-                  <td><font color="black">'.$cat_name.'</font></td> 
-				  <td><a href="blog-content.php?i='.$field1name.'">Content</td>
-				  <td><a href="edit-ipodetail.php?i='.$field1name.'">IPO Detail</td>
-				  <td><font color="black">'.$publish.'</font></td>
-				  <td><font color="black">'.$publish_date.'</font></td>
-                  <td><a href="edit-blog.php?i='.$field1name.'">Edit</td>';
-				  if($_SESSION['user']['username'] == 'finoadmin')
-				  {	  
-				  echo '<td><a href="delete-blog.php?i='.$field1name.'">Delete</td>';
-				  }
-            echo '</tr>';
-        }		 
     }
     $result->free();
 } 
+echo '</table>';
+?>
 
-
-			?>
+<div class="pagination-container" style="width: 70%; margin: 20px auto; text-align: center;">
+    <div class="btn-group">
+        <?php if($page > 1): ?>
+            <a href="?page=1&search=<?php echo urlencode($search); ?>" class="btn btn-outline-primary" style="border: 1px solid #3f6791; color: #3f6791;">First</a>
+            <a href="?page=<?php echo $page-1; ?>&search=<?php echo urlencode($search); ?>" class="btn btn-outline-primary" style="border: 1px solid #3f6791; color: #3f6791;">Previous</a>
+        <?php endif; ?>
+        
+        <?php
+        $start_loop = max(1, $page - 2);
+        $end_loop = min($total_pages, $page + 2);
+        
+        for($i = $start_loop; $i <= $end_loop; $i++): ?>
+            <a href="?page=<?php echo $i; ?>&search=<?php echo urlencode($search); ?>" class="btn <?php echo ($i == $page) ? 'btn-primary' : 'btn-outline-primary'; ?>" style="<?php echo ($i == $page) ? 'background: #3f6791; border: 1px solid #3f6791;' : 'border: 1px solid #3f6791; color: #3f6791;'; ?>"><?php echo $i; ?></a>
+        <?php endfor; ?>
+        
+        <?php if($page < $total_pages): ?>
+            <a href="?page=<?php echo $page+1; ?>&search=<?php echo urlencode($search); ?>" class="btn btn-outline-primary" style="border: 1px solid #3f6791; color: #3f6791;">Next</a>
+            <a href="?page=<?php echo $total_pages; ?>&search=<?php echo urlencode($search); ?>" class="btn btn-outline-primary" style="border: 1px solid #3f6791; color: #3f6791;">Last</a>
+        <?php endif; ?>
+    </div>
+    <div style="margin-top: 10px; font-family: poppins;">
+        Page <?php echo $page; ?> of <?php echo $total_pages; ?> (Total <?php echo $total_rows; ?> blogs)
+    </div>
+</div>
 
 	
 	 <style>
@@ -184,16 +213,12 @@ if ($result = mysqli_query($link, $query)) {
     font-size: 15px;
     padding: 7px;
 }
-	 }
 	 
 	 </style>
 	 
         <!-- /.row -->
       </div><!--/. container-fluid -->
     </section>
-	
-	
-	
 	
 
     <!-- /.content -->
